% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_duration}
\alias{calc_duration}
\title{calc_duration}
\usage{
calc_duration(start = NULL, end = NULL, units = "years")
}
\arguments{
\item{start}{Required. Date or Timestamp. The start date.}

\item{end}{Required. Date or Timestamp. The end date.}

\item{units}{Optional. Character. Units of the returned duration
(i.e. 'seconds', 'days', 'years'). Defaults to 'years'.}
}
\description{
Returns the duration of time between two provided date objects.
}
