% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{paste_mean}
\alias{paste_mean}
\title{Paste mean}
\usage{
paste_mean(x, less.than.one = FALSE, digits = 1)
}
\arguments{
\item{x}{A numeric. Data to summarize.}

\item{less.than.one}{A logical. Indicates a mean that rounds to 0 should
be printed as <1.}

\item{digits}{An integer. Number of digits to round to.}
}
\value{
A character vector of the mean(s) with standard deviation(s).
}
\description{
Creates a human-readable mean with standard deviation
from numeric data.
}
\examples{
paste_mean(mtcars$mpg)
}
