% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{paste_pval}
\alias{paste_pval}
\title{Paste p-value}
\usage{
paste_pval(x, digits = 1, p.digits = 4)
}
\arguments{
\item{x}{A numeric. P-value to format.}

\item{digits}{A numeric. Number of significant digits to round to.}

\item{p.digits}{A numeric. Minimum number of digits to right of the decimal
point.}
}
\description{
Creates a human-readable p.value using sensible defaults for `format.pval()`.
}
\examples{
paste_pval(0.061126e-10)
}
