% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{paste_text}
\alias{paste_text}
\alias{paste0_text}
\title{Paste Text}
\usage{
paste_text(..., na.rm = TRUE)

paste0_text(..., na.rm = TRUE)
}
\arguments{
\item{...}{Required. R objects to be converted to character vectors or
documented 'base::paste()' or 'base::paste0()' parameters.}

\item{na.rm}{Optional. Logical. Whether to remove NA values from 'x'.
Note that NA values are also removed from vectors. Defaults
to TRUE.}
}
\value{
Character vector of concatenated values.
}
\description{
An augmented 'base::paste()' with optional management
of NA values.
}
\details{
The 'base::paste()' function is intentionally designed to
coarce NA values to characters that appear in the concatenated
character output. However, this behavior is not always desirable (i.e.
when the input is a dynamically generated vector of character strings
that contains some NA values) and there is currently no means of opting
out of this behavior.
}
\examples{
# Default 'base::paste()' handling of NA values
paste_text(
  'The', c('red', NA, 'orange'), 'fox jumped', NA, 'over the fence.',
  collapse = ' ',
  na.rm = FALSE
)

# Removal of NA values
paste_text(
  'The', c('red', NA, 'orange'), 'fox jumped', NA, 'over the fence.',
  collapse = ' '
)
}
