% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate_coef.glm}
\alias{tabulate_coef.glm}
\title{Tabulate glm coefficients}
\usage{
\method{tabulate_coef}{glm}(
  fit,
  level = 0.95,
  test = c("LRT", "Chisq", "Rao"),
  format = FALSE,
  digits = 1,
  p.digits = 4
)
}
\arguments{
\item{fit}{Required. An object of class '\code{\link[stats]{glm}}'.}

\item{level}{Optional. Double. Confidence level.}

\item{test}{Optional. Character. Test to use for stats::drop1 testing of
model parameters.}

\item{format}{Optional. Logical. Round numbers and format text for a
cleaner, readable output.}

\item{digits}{Optional. Integer. Number of digits to round to.}

\item{p.digits}{Optional. Integer. Number of p-value digits to print. Note that
p-values are still rounded according to 'digits' value.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} of model
terms with the columns:
\item{variable}{The name of regression variable.}
\item{level}{The level of the regression varaible, if the
variable is a factor.}
\item{observations}{The number of observations used in the fit.}
\item{events}{The number of events used in the fit.}
\item{estimate}{The odds ratio of the regression term.}
\item{conf.low}{The low end of the confidence interval for the
regression term odds ratio.}
\item{conf.high}{The low end of the confidence interval for the
regression term odds ratio.}
\item{test}{The test used to produce the statistic and p-value.}
\item{statistic}{The value of a statistic used in testing the null
hypothesis.}
\item{p}{The two-sided p-value associated with the statistic.}
}
\description{
Converts terms from a generalized linear model
into a usable table for publication purposes.

This function is fundamentally similar to broom::tidy with
some important differences that enable creation of more human-readable
outputs (i.e. exportable tables). All model variables and their respective
levels, not simply model terms, are reported including the reference level
for factors with 3 or more levels. A test statistic is also calculated for
each factor variable with multiple "dummy variables". Variable statistics
can also be overwritten with model \code{\link[stats]{drop1}} test
statistics. Methods are also provided for the return of a fully-formatted
export-ready table.
}
\examples{
library(dplyr)
library(MASS)

logit_data <- MASS::birthwt \%>\%
  mutate_at(c('race'), as.factor) \%>\%
  mutate_at(c('low', 'smoke', 'ht', 'ui'), as.logical)

fit_glm <- stats::glm(
  low ~ race + smoke + age,
  data = logit_data,
  family = 'binomial')

tabulate_coef(fit_glm) # Format ready
tabulate_coef(fit_glm, format = TRUE) # Formatted
}
\seealso{
\code{\link{tabulate_coef}}
}
