% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate_coef}
\alias{tabulate_coef}
\alias{tabulate_coefficients}
\title{Tabulate model coefficients}
\usage{
tabulate_coef(fit, level, test, format, digits, p.digits)

tabulate_coefficients(fit, level, test, format, digits, p.digits)
}
\arguments{
\item{fit}{Required. Model object. See S3 methods below.}

\item{level}{Optional. Double. Confidence level.}

\item{test}{Optional. Character. Test to use for stats::drop1 testing of
model parameters.}

\item{format}{Optional. Logical. Round numbers and format text for a
cleaner, readable output.}

\item{digits}{Optional. Integer. Number of digits to round to.}

\item{p.digits}{Optional. Integer. Number of p-value digits to print. Note that
p-values are still rounded according to 'digits' value.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} containing summarizing statistics and tests.
}
\description{
Converts parameters from a model object into a usable
table for publication purposes.

This function is fundamentally similar to broom::tidy with
some important differences that enable creation of more human-readable
outputs (i.e. exportable tables). All model variables and their respective
levels, not simply model terms, are reported including the reference level
for factors with 3 or more levels. A test statistic is also calculated for
each factor variable with multiple "dummy variables". Variable statistics
can also be overwritten with model \code{\link[stats:add1]{stats::drop1}} test
statistics. Methods are also provided for the return of a fully-formatted
export-ready table.
}
\seealso{
\code{\link{tabulate_coef.lm}},
\code{\link{tabulate_coef.coxph}},
\code{\link{tabulate_coef.glm}}
}
