% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate_logit}
\alias{tabulate_logit}
\title{tabulate_logit}
\usage{
tabulate_logit(formula = NULL, data = NULL)
}
\arguments{
\item{formula}{Required. Formula. Y ~ X1 + X2 + X3}

\item{data}{Required. Tibble. Data used by the formula.}
}
\value{
Returns model object containing:
  \item{obsCount}{The model's observation count.}
  \item{paramCount}{The model's parameter count.}
  \item{AIC}{The model's AIC.}
  \item{AICc}{The model's AICc.}
  \item{parameters}{A data.frame (tibble) of the model's estimated parameter coefficients.}
  \item{ORs}{A data.frame (tibble) of model parameters' odds ratios and confidence intervals calculated using Likelihood Ratio Tests.}
  \item{LRTs}{A data.frame (tibble) of Likelihood Ratio Test results for each model parameter.}
  \item{model}{The logistical regression model object.}
}
\description{
Legacy function. Returns list of summary statistics from
data for a logistic regression model. Functionality will eventually be replicated
by tabulate_model().
}
