% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_hypothesis.logical}
\alias{test_hypothesis.logical}
\title{Test Hypothesis: Logical}
\usage{
\method{test_hypothesis}{logical}(
  x,
  y,
  test = c("chisq", "fisher"),
  digits = 1,
  p.digits,
  simulate.p.value = FALSE,
  B = 2000,
  workspace = 2e+07,
  ...
)
}
\arguments{
\item{x}{A logical. Observations.}

\item{y}{A factor or logical. Categorical "by" grouping variable.}

\item{test}{A character. Name of the statistical test to use. Supported tests
include Pearson's Chi-squared Test ('chisq') and Fisher's Exact Test ('fisher').}

\item{digits}{An integer. Number of digits to round to.}

\item{p.digits}{An integer. Minimum number of p-value digits to the right of
the decimal point. Note that p-values are still rounded using 'digits'.}

\item{simulate.p.value}{A logical. Whether p-values in nominal variable testing
should be computed with Monte Carlo simulation.}

\item{B}{An integer. Number of replicates to use in Monte Carlo simulation.}

\item{workspace}{An integer. Size of the workspace used for the Fisher's Exact
Test network algorhythm.}

\item{...}{Additional arguments passed to S3 method.}
}
\value{
A list containing the statistical test performed, test statistic,
and p-value.
}
\description{
Test Hypothesis: Logical
}
\seealso{
\code{\link{test_hypothesis}}
}
