% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_hypothesis.numeric}
\alias{test_hypothesis.numeric}
\title{Test Hypothesis: Numeric}
\usage{
\method{test_hypothesis}{numeric}(
  x,
  y,
  test = c("anova", "kruskal", "wilcoxon"),
  digits = 1,
  p.digits,
  ...
)
}
\arguments{
\item{x}{A numeric. Observations.}

\item{y}{A factor or logical. Categorical "by" grouping variable.}

\item{test}{A character. Name of the statistical test to use. Supported tests
include ANOVA linear model ('anova'), kruskal-wallis ('kruskal'), and wilcoxon
rank sum tests ('wilcoxon').}

\item{digits}{An integer. Number of digits to round to.}

\item{p.digits}{An integer. Minimum number of p-value digits to the right of
the decimal point. Note that p-values are still rounded using 'digits'.}

\item{...}{Additional arguments passed to S3 method.}
}
\value{
A list containing the statistical test performed, test statistic,
and p-value.
}
\description{
Test Hypothesis: Numeric
}
\seealso{
\code{\link{test_hypothesis}}
}
