% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_hypothesis}
\alias{test_hypothesis}
\title{Test the null hypothesis}
\usage{
test_hypothesis(x, y, test, digits, p.digits, ...)
}
\arguments{
\item{x}{A numeric, factor, or logical. Observations.}

\item{y}{A factor or logical. Categorical "by" grouping variable.}

\item{test}{A character. Name of the statistical test to use. See note.}

\item{digits}{An integer. Number of digits to round to.}

\item{p.digits}{An integer. Minimum number of p-value digits to the right of
the decimal point. Note that p-values are still rounded using 'digits'.}

\item{...}{Additional arguments passed to the appropriate S3 method.}
}
\value{
A list containing the statistical test performed, test statistic,
and p-value.
}
\description{
Tests the null hypothesis that there is no difference between
grouped data.
}
\details{
Statistical testing used is dependent on type of 'x' data. Supported
testing for numeric data includes ANOVA ('anova'), Kruskal-Wallis ('kruskal'),
and Wilcoxon Rank Sum ('wilcoxon') tests. For categorical data, supported
testings includes Pearson's Chi-squared ('chisq') and Fisher's Exact Test
('fisher').
}
\examples{
strata <- as.factor(mtcars$cyl)

test_hypothesis(mtcars$mpg, strata) # numeric
test_hypothesis(as.logical(mtcars$vs), strata) # logical
test_hypothesis(as.factor(mtcars$carb), strata) # factor
}
\seealso{
\code{\link{test_hypothesis.numeric}},
\code{\link{test_hypothesis.factor}},
\code{\link{test_hypothesis.logical}}
}
