% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_numeric}
\alias{test_numeric}
\title{Test Numeric Data}
\usage{
test_numeric(col = NULL, by = ".by", data = NULL,
  parametric = FALSE, digits = 1, p.digits = 4)
}
\arguments{
\item{col}{Required. Character. Name of numeric column containing observations.}

\item{by}{Required. Character. Name of logical or factor column to stratify by.}

\item{data}{Required. Tibble. Data being used.}

\item{parametric}{Optional. Logical. Indicates parametric testing should be used.
(Student's T-Test). Defaults to FALSE (non-parametric; Wilcox).}

\item{digits}{Optional. Integer. Number of digits to round to. Defaults to 1.}

\item{p.digits}{Optional. Integer. Number of p-value digits to print. Note that
p-values are still rounded using 'digits'. Defaults to 4.}
}
\description{
Returns p-value from parametric or non-parametric testing of stratified
continuous (numeric) data.
}
