% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk_data_}
\alias{chunk_data_}
\title{Break a data object into chunks}
\usage{
chunk_data_(data = NULL, size = 10, reverse = FALSE)
}
\arguments{
\item{data}{Required. Tibble, data frame, vector.}

\item{size}{Optional. Integer. The number of items (e.g. rows in a tibble)
that make up a given chunk. Must be a positive integer. Defaults to 10.}

\item{reverse}{Optional. Logical. Calculate chunks from back to front.
Defaults to front to back ordering.}
}
\value{
A factory function which returns a chunk of data from the provided
object with each call. Once all data has been returned, function returns
NULL perpetually.
}
\description{
Creates a factory function which returns a different chunk
of a given data object with each function call.
}
\examples{
# Create chunk factory function
chunked_data <- chunk_data_(mtcars, size = 6)

# Chunk #1 (rows 1-6)
paste0(rownames(chunked_data()), collapse = ', ')

# Chunk #2 (rows 7-12)
paste0(rownames(chunked_data()), collapse = ', ')
}
