% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{paste_median}
\alias{paste_median}
\title{Paste Median}
\usage{
paste_median(x = NA, less.than.one = FALSE, digits = 1)
}
\arguments{
\item{x}{Required. Numeric. Data to summarize.}

\item{less.than.one}{Optional. Logical. Indicates a median that rounds to 0 should
be printed as <1. Defaults to FALSE (0).}

\item{digits}{Optional. Integer. Number of digits to round to. Defaults to 1.}
}
\description{
Returns a human-readable median with inter-quartile
range from numeric data.
}
\examples{
paste_median(mtcars$mpg)
}
