% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate_model}
\alias{tabulate_model}
\title{Tabulate Model}
\usage{
tabulate_model(fit, format, percent.sign, digits, p.digits)
}
\arguments{
\item{fit}{Required. Model object. See S3 methods below.}

\item{format}{Optional. Logical. Round numbers and format text for a
cleaner, readable output.}

\item{percent.sign}{Optional. Logical. Print percent sign after frequencies.}

\item{digits}{Optional. Integer. Number of digits to round to.}

\item{p.digits}{Optional. Integer. Number of p-value digits to print. Note that
p-values are still rounded according to 'digits' value.}
}
\value{
Returns tibble containing summarizing statistics and tests.
}
\description{
Converts parameters from a model object into a usable
table for publication purposes.
}
\seealso{
\code{\link{tabulate_model.lm}},
\code{\link{tabulate_model.coxph}},
\code{\link{tabulate_model.glm}}
}
